add parser for wikitext links
