+++ package.json
@@ -12,15 +12,19 @@
   "license": "MIT",
   "dependencies": {
     "axios": "^0.21.1",
+    "bufferutil": "^4.0.2",
     "cron": "^1.8.2",
     "date-fns": "^2.16.1",
     "discord-akairo": "^8.1.0",
     "discord.js": "^12.5.1",
     "dotenv": "^8.2.0",
+    "erlpack": "github:discord/erlpack",
     "esm": "^3.2.25",
     "rollbar": "^2.19.4",
     "signale": "^1.4.0",
-    "string-similarity-js": "^2.1.2"
+    "string-similarity-js": "^2.1.2",
+    "utf-8-validate": "^5.0.3",
+    "zlib-sync": "^0.1.7"
   },
   "optionalDependencies": {
     "bufferutil": "^4.0.2",
